Update settings.py anonymous actions
This relates to a security audit recommendation to remove anonymous action like 'like' to occur when anonymous users are in the Media Page. 

With these changes, a pop-up message 'Action failed' is produced. 

Note: A better fix would be to divert anonymous users to log-in when they click the like button.